refactor(services): extract scorecard-builder and standings modules
Verbatim moves of the four module-level pure functions out of
score-service.ts (spec 003 Group 4, DD-3): buildSeasonData +
buildScorecardTeamBlock to scorecard-builder.ts; computeStandings +
recomputeStandingsFromWeeks to standings.ts, whose header records the
two-derivations invariant and the DD-4 unification deferral. Both new
modules import only types + scoring-engine. All 236 unit tests pass
unmodified.
